requirements.txt
pypng==0.0.20
PyQRCode==1.2.1
PySimpleGUI==4.34.0
selenium==3.141.0
urllib3==1.26.2
